Refactor user API boilerplate into separate api_* prefixed files
